Move main to test
